refactor codigo duplicado,nova classe abstrata para implementacao
